requirements.txt
# 核心依赖
Flask>=2.0.0
Flask-Cors>=3.0.10
opencv-python>=4.5.5.64
Pillow>=9.0.0
numpy>=1.21.0
requests>=2.27.0
beautifulsoup4>=4.11.0
selenium>=4.1.0
werkzeug>=2.0.0

# 可选依赖 - 用于深度学习（如果需要更高级的图像分析）
tensorflow>=2.8.0
torch>=1.11.0
torchvision>=0.12.0

# 开发和生产环境
gunicorn>=20.1.0
python-dotenv>=0.20.0

# 用于图像处理的额外库
scikit-image>=0.19.0
matplotlib>=3.5.0
